requirements.txt
APScheduler
Pyrogram
TgCrypto
Deta
bs4
requests
python-dotenv
flask
gunicorn
psutil